SnpEff analysis for determining SNPs and Indels
In this section of the shell script, the vcf files generated from the freebayes pipeline are fed. From the vcf files, SNPs and Indels are separated using vcftools. The common SNPs and Indels in the nonirradiated parental line are removed using the vcf-isec command to get a normalized vcf file. These variants are then analysed using the SnpEff tool to obtain statistical data.